Correct GC suspension timing and activity correlation (#953)
Separate GC collection elapsed from versioned fully-suspended evidence.
Bound event pairing, preserve evidence quality, and attribute suspension
using per-activity interval union through shared CLI/MCP query paths.

Closes #950.
Separate follow-ups remain tracked in #940 and #952.

+++ src/DotnetDiagnostics.Mcp/Tools/CollectBatchTool.cs
@@ -555,7 +555,7 @@ public sealed record CollectBatchGen2Evidence(
     double? MeterRatePerSecond,
     double? MeterProcessCumulative,
     int GcCollectorWindowCount,
-    int GcCollectorWindowSeconds,
+    double GcCollectorWindowSeconds,
     string Explanation);
 
 /// <summary>
